Add git commit.vim
This file will make the gitcommits automaticallly break lines at 50 for
the title, and 72 for the commit msg.

I prefer this way over having som highlightning that shows when my lines
are too long.

If I need a longer line, it's still possible to do that, and when
pasting like a long URL - it'll still paste as a long URL, without
a linebreak.

Note: Would really like to find out how to make the first line (the
summary) to add an xtra line break, so it jumps to line 3.