Upd .bashrc to current version
Not really much have changed. Just some re-ordering of things and is the
one I'm currently using in Rocky. So, I know it works on both now.

Added a function near the end, that is for readline, and binds C-l to
clear the screen - incl the output buffer. It's not perfect, but it
works as intended most of the times (can leave previous cuts into the
re-paste of the copied string, if empty)